v0.1 - Controlling the Character
 -- Overall --
As the Player needs to be controlled by an AI Controller to use the MoveToLocation function, I had to make a large workaround.  The player controls the Player APawn, which is attached to the Person ACharacter which is controlled by a AIController
Additionally, I worked with PlayerControllers for inputs over directly implementing them in the Pawn/Character class.
Finally, I learned how to use the new EnhancedInput system, due to the previous system being deprecated

-- Prototyping --
Protoyped Controllers, Character, AI and Gamemode in Blueprint
 - Tagged with the PROTO_ name

-- SingleInputAIControlled --
Added to project, used by SingleInputPerson

-- SingleInputPlayer --
Added the CharacterCamera component

-- SingleInputPlayerController --
Added pointers to the OwnedPawn (SingleInputPlayer), AICharacter (SingleInputPerson) and AIControl (SingleInputAIController)
Added pointers to the input data objects
Added debug boolean, allowing use for debug funtions

Implemented ConsrtuctorHelpers to find and store the input data objects
Implemented character setup in BeginPlay
- Starts by casting and storing to the owned pawn
- Next, spawns the AICharacter, then spawns the AIController and make the AIController possess the Character
- Then attach the Player to the Persons spring arm and make the cursor visible on screen.

Implemented the Enhanced Input System
- Binded PlayerInteract, RotateCamera and AngleCamera
Implemented PlayerInteract, which gets the mouse position and direction and calls the MoveToLocation function in Person
Implemented RotateCamera and AngleCamera Debug functions
- RotateCamera detects if the camera should be rotated clockwise or counter-clockwise, then calls the RotateCameraByStep function in Person
- AngleCamera calls the SwapCameraAngle function in Person

-- SingleInputPerson --
Added float properties to control the camera
Added pointers to the AIController
Added SpringArm component

Implemented SwapCameraAngle function, which changes the Pitch of the spring arm to the opposite value when called
Implemented RotateCameraByStep function, which adds a step of rotation on the spring arms Yaw based on the direction inputted.
Implemented MoveToLocation, which takes in a location and a direction to fire a line trace to find a location to move to, then calls the MoveToLocation function in the AIController.

 -- Inputs --
Added InputConfigData DataObject, which stores all of the InputActions and InputMappingContext assets together.