v0.11 - Whole System Rework
Reworked and Remade the whole project up to v0.10

Problems started to arise due to the structure of the project.  I renamed and reordered most classes into seperate folders to better denote what each class does, instead of everything being in Core

I also simplified how each component of the player can communicate with each other, removing erroneous pointers and improving the structure entirely
For example, the PlayerController no longer calls a function in PlayerCharacter to move the player, instead going straight to the AIController directly.  Additionally, the UI structure has been updated to simply functions

A new addition is the ItemManager, which holds the Item/Recipe data tables instead of the inventory components.  A template is used to get data from the item data tables instead of multiple functions.  Additionally, the ItemDisplay is now contained in the ItemManager instead of a seperate object spawned in the world.

The inventory no longer stores the whole item data, instead just an item ID, the amount and any additional information unique to that item (inventory order and upgrades)
Most of the code is the same or similar, with some properties names being changed to better suit how they are used.

With how the project is setup now, additions such as weapon stats and consumable data should be easier to implement