v0.13 - Storage and Stats
 -- ParentStorage --
Added subclass of ParentInteractable
Has its own InventoryComponent, allowing users to store items inside
When UseInteractable is called, a new state is shown (StorageState) allowing players to transfer items between their inventory and the storage object

 -- ParentInteractable --
Added parent class of interactable world objects
Added overridable UseInteractable function

-- ParentCraftingStation --
Now a subclass of ParentInteractable
Overrieds UseInteractable, when called opens the CraftingState if the user is a player

 -- StorageState --
Subclass of ParentState
Displays two separate TileViews of both the players inventory and the storages inventory
Works in a similar way to InventoryState (see Images/13-001.png to Images/13-002.png)
Players can sort each inventory separatley
Allows players to move either a single item or a stack of items at a time, based on if they left or right click a StorageSlot, through MoveItemSingle or MoveItemStack
Will unequip weapons when transferring them to a storage inventory if equipped

-- StorageSlot --
Looks identical to InventorySlot
Now displays the amount of the item as a text box on the item image
If the button is left clicked, MoveItemSingle is called
If the button is right clicked, MoveItemStack is called
The slot has a boolean to state if it is in the Player or Storage list
See Images/v13-003.png to Images/v013-006.png

 -- PlayerUI --
Added StorageState and updated the widget switcher functions to allow us of the state

 -- InventorySlot --
Added TextBlock that displays the amount of the item in the inventory

 -- InGameState --
Updated UpdateHealthBar to now use CurrentHealth and MaxHealth arguments, and set the HealthBar to the percentage of health remaining

 -- WeaponItem --
Implemented ClearWeapon, which empties the WeaponData structs and clears the skeletal mesh

 -- PlayerController --
Moved GetPlayerUI from protected to public
Added GetAICharacter, which simply returns a pointer to the AICharacter
Replaced IsA(AParentCraftingStation) with IsA(AParentInteractable), removed unneeded Cast and updated SetActiveStateToStation to return TraceHit.GetActor()
Now passes the PlayerUI to the StatsComponent and calls SetupStats

-- PlayerCharacter --
Implemented UnequipWeapon, which simply calls ClearWeapon on the players weapon
Added StatsComponent
Now stores an integar of the equipped weapons order in the inventory

 -- AICharacter --
Reimplemented SetPCBlackboardValue
Updated SetActiveStateToStation to use AActor instead of AParentCraftingStation as an argument

 -- StatsComponent --
Contains all of the stats of a player, enemy or damageable object
Implemenetd TakeDamage function, which calculates the amount of damage a object should take based on its Defence
Implemented HealUnit, which is used to heal the unit
Added GetCurrentHealth and GetMaxHealth, which simply return the stats

 -- InventoryComponent --
When removing, now checks if the item to remove is a weapon.  If it is, check if the component owner is a player.  If so, now casts to the ASI_PlayerCharacter class and stores the pointer
When querying the inventory, now checks if the pointer is set (if true, the item is a weapon), followed by a check to see if the weapon's order matches the EquippedWeaponOrder of the player.  If it does, the weapon is unequipped before storing/removing.
Now adds the inventory back on removal

 -- UI --
Added BP classes of StorageSlot and StorageState, as well as a StorageSlotData class
See Images/v013-00x.png for functions

 -- AI --
Updated UseStation task to use the new ParentInteractable function