v0.14 - Buff System and Consumables
 -- InventoryState --
Added a string to store the current selected item's name
Updated OnSelectedItemUseButtonRelease to update the description box when an item is used instead of needing to re-set the CurrentSelectedItem each time.  Additionally clears the Description Box and removes the selected item from the ListView if all the items are used.
Now checks if the selected item is either a weapon or a consumable when deciding to show the use button

 -- InventorySlot --
Now stores a pointer to the slot's item data object

 -- InGameState --
UpdateHealthBar now uses float arguments instead of integars
Now also shows the player's health as a text box as well as a progress bar via PlayerHealthText

 -- ItemManager --
Implemented GetConsumableDataFromID, as well as a pointer to the ConsumableConst data table

 -- ItemData --
Changed how EConsumableType is used.
 - Was used to denote what stat is modified by the consumable
 - Now denotes what buff type is used by the consumbale
Updated the FConsumableData struct with updated properties

 -- StatsComponent --
Added TMap ActiveBuffMap, used to store data about active buffs affecting the owner of the StatsComponent
Implemented AddBuff and RemoveBuff, used to add or remove buffs from the TMap
Implemented GetStatByTag, which returns the value of a stat
Implemented UpdateStat, which modifies the value of a stat
Implemented StartConsumableBuff, which is called to begin a buff from a consumable only
 - Other buffs will use the StartBuff function instead
 - This is due to consumables using different stat amounts and stat types
 - Can also be used to refresh a current consumable buff

 -- InventoryComponent --
UseItem now checks if a item has a type of Consumable.  If it is, StartConsumableBuff is called and the item is removed from the inventory

 -- ParentBuffDebuff --
Added super class for the buff system
 - Added virtual functions for StartBuff, RefreshBuff, BuffTickTimer and EndBuff
 - Added BuffName property and UStatComponent pointer

 -- ParentBuff / ParentDebuff --
Super classes for each buff or debuff

 -- StatBuff --
Buff subclass used to increase a players stat
Implemented SetupBuff, which is used to setup the buff to the consumables stats.
 - Takes in StatTag, BaseStat, Multiplier and Duration.
 - When the buff is started, a duration timer is set and the buff is added to the attatchee's stats component
 - Next, the total stat change is calculated and stored
 - Finally, the stat change is updated
Implemented SetupReBuff, which is used to change the stats of the buff to new ones
 - Removes the need to create a new object each time a new buff of the same stat is created

 -- RegenBuff --
Implemented SetupBuff, which acts in a similar way to StatBuff
 - Also addes a TickTimer and uses the BuffTick function
 - Each time the BuffTick function is called, the StatsComponent is healed.

 -- Blueprint UI --
Added the PlayerHealthText to InGameState
InventorySlot now stores the data object before casting to it
Updated ItemConst and ConsumableConst with consumable items