v0.15 - Combat
 -- StorageState --
Now displays the name of the storage in the UI

 -- InventoryState --
Now stores the selected item's type, which is then queried when the Use button is pressed
Weapon items are checked to see if they have a max stack of 1 before the use button is displayed
Removed refresh description box

 -- InGameState --
Now displays the icon and total ammo count for the equipped weapon (see Images/v15-001.png)

 -- WeaponItem --
Added fire rate functionality through a boolean and a timer
 - If the boolean is true, the weapon can fire
 - Fire delay is calculated via FireRate / 60
Added functions for charging the weapon's accuracy to increase probability of the shot hitting
 - Starts a timer that ticks every 0.5 seconds, increasing the accuracy of the weapon
 -- Increase per tick is calculated via MinAccuracy - MaxAccuracy / 0.5s
 - Timer is ended when MaxAccuracy is reached

 -- ItemManager --
Added slight rotation on the item meshed

 -- ItemDatat --
Added ENUM_RANGE_BY_COUNT MACRO to ItemType
Added CritChance, CritDamage, AccuracyChargeTime, MinAccuracy and MaxAccuracy to FWeaponData
Changed BaseStat from an int to a float in FConsumableData

-- PlayerController --
Now passes the PlayerUI to the InventoryComponent on BeginPlay
When comparing what the InteractTrace has hit, now checks if the actor has a StatsComponent.  If they do, their team is then compared.  If so, the active state of the PlayerAI is set to Combat.

 -- PlayerCharacter --
Now un-equips the weapon before re-equipping a new one in EquipWeapon
Also both the StatsComponent and InventoryComponent's are updated when a weapon is equipped.
On weapon unequip, the StatsComponent is updated again

 -- AIController --
Added CombatState function

 -- StatsComponent --
Updated TakeDamage to now have arguments for CritChance and CritMultiplier
 - Calculates if an input is a crit via FMath::RandRange(0.0f, 1.0f) <= CritChance
Removed BonusDefence property
Now checks if the CurrentHeath drops below 0 when damaged.  If so, the owner of the StatsComponent is destoryed
HealUnit now updates the player's health bar if the StatsComponent is owned by a player
Updated GetStatByTag to now return all combat stats
Implemneted GetTeam, which returns the ObjectTeam integar
Removed GetUnitBonusDamage and GetUnitDamageMultiplier functions as they now can be accessed via GetStatByTag
UpdateStat now updates the player's health bar if the StatsComponent is owned by a player and the stat to update is Health
Added all combat stats to UpdateStat
Added UpdateCombatStatsFromWeapon, which updates the combat stats via FWeaponData
 - Also updates the weapon icon on InGameState

 -- InventoryComponent --
Now stores the AmmoID of the equipped weapon
Also updates the ammo count on the InGameState when AddItemToInventory or RemoveItemFromInventory is called, if the ItemID matches the AmmoID
Added a pointer to the PlayerUI, only used if the owner of the component is a player
Implemented GetAmountOfItem, which returns the amount of a specific ItemID exists in the inventory
Now initalizes the inventory map with empty arrays if the array doesn't already exist via a for loop on construction, done vis TEnumRange and ENUM_RANGE_BY_COUNT

 -- Prototyping --
Added Target prototype asset, used to test the combat state
Added HitTarget BT Task, used to target and damage the PROTO_Target

 -- TestEnvironment --
Added assets to the test environment

 -- Icon --
Updated default parameter for the material instance

 -- InGameState BP --
Added Image and TextBlock assets for the weapon and ammo count

 --Data Tables --
Added new consumables, updated TestRifle Weapon stats

 -- AI --
Added MoveInRangeOfTarget (see Images/v15-002.png)
Added GetUnitRange (v15-003)
Added FireWeaponAtTarget(v15-004 - v15-007)
Added StartWeaponAccuraty(v15-008)
Updated BehaviorTree with the combat state (v15-009)